scheduling_service/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==4.2.16
djangorestframework==3.15.2
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-restful==0.6.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
kafka-python==2.0.2
lxml==5.3.0
mongoengine==0.29.1
mypy-extensions==1.0.0
packaging==24.1
psutil==5.9.8
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-decouple==3.8
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
starlette==0.41.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
